Agregando funcionalidades a la compra

+++ unicine/unicine_front/lib/models/unicine/purchase.dart
@@ -16,8 +16,8 @@ class Purchase {
   int? idCompra;
   DateTime? fechaCompra;
   String? metodoPago;
-  int? subtotal;
-  int? total;
+  double? subtotal;
+  double? total;
 
   factory Purchase.fromJson(String str) => Purchase.fromMap(json.decode(str));
